Implementing ExceptionHandling

+++ PageAppWeb/Services/StudentService.cs
@@ -1,5 +1,6 @@
 ﻿using PageApp.Infrastracture.Models;
 using PageApp.Infrastracture.Repositories;
+using PageAppWeb.Exceptions;
 
 namespace PageAppWeb.Services;
 
@@ -20,6 +21,10 @@ public class StudentService : IStudentService
     public async Task DeleteStudent(int id)
     {
         var student = await _studentRepository.GetById(id);
+
+        if (student == null)
+            throw new StudentNotFoundException(id);
+
         await _studentRepository.Delete(student);
     }
 
@@ -32,14 +37,14 @@ public class StudentService : IStudentService
     {
         var studentToBeUpdated = await _studentRepository.GetById(id);
 
-        if (studentToBeUpdated != null)
-        {
-            studentToBeUpdated.Name = student.Name;
-            studentToBeUpdated.Surname = student.Surname;
-            studentToBeUpdated.IndexNumber = student.IndexNumber;
-            studentToBeUpdated.StudentStatusId = student.StudentStatusId;
-            studentToBeUpdated.Year = student.Year;
-        }
+        if (studentToBeUpdated == null)
+            throw new StudentNotFoundException(id);
+
+        studentToBeUpdated.Name = student.Name;
+        studentToBeUpdated.Surname = student.Surname;
+        studentToBeUpdated.IndexNumber = student.IndexNumber;
+        studentToBeUpdated.StudentStatusId = student.StudentStatusId;
+        studentToBeUpdated.Year = student.Year;
 
         return await _studentRepository.Update(studentToBeUpdated);
     }
